requirements.txt
streamlit==1.39.0
requests==2.32.3
pydantic==2.11.0
langchain==1.0.3
langchain-core==1.0.2
langchain-ollama==1.0.0
langchain-groq==0.0.7
openai==0.28.0
langgraph==0.4.0
python-dotenv==1.0.1

